DEFEDIT-1587 Cancel bundling v 1.0

+++ com.dynamo.cr/com.dynamo.cr.bob/src/com/dynamo/bob/bundle/IOSBundler.java
@@ -109,9 +109,13 @@ public class IOSBundler implements IBundler {
     }
 
     @Override
-    public void bundleApplication(Project project, File bundleDir)
+    public void bundleApplication(Project project, File bundleDir, ICanceled canceled)
             throws IOException, CompileExceptionError {
         logger.log(Level.INFO, "Entering IOSBundler.bundleApplication()");
+
+        if(canceled.isCanceled()) {
+            return;
+        }
 
         // Collect bundle/package resources to be included in .App directory
         Map<String, IResource> bundleResources = ExtenderUtil.collectResources(project, Platform.Arm64Darwin);
@@ -125,6 +129,9 @@ public class IOSBundler implements IBundler {
         // If a custom engine was built we need to copy it
         String extenderExeDir = FilenameUtils.concat(project.getRootDirectory(), "build");
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // armv7 exe
         {
             Platform targetPlatform = Platform.Armv7Darwin;
@@ -141,6 +148,9 @@ public class IOSBundler implements IBundler {
         }
         File exeArmv7 = binsArmv7.get(0);
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // arm64 exe
         {
             Platform targetPlatform = Platform.Arm64Darwin;
@@ -160,6 +170,9 @@ public class IOSBundler implements IBundler {
         logger.log(Level.INFO, "Armv7 exe: " + getFileDescription(exeArmv7));
         logger.log(Level.INFO, "Arm64 exe: " + getFileDescription(exeArm64));
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         BobProjectProperties projectProperties = project.getProjectProperties();
         String title = projectProperties.getStringValue("project", "title", "Unnamed");
         String exeName = BundleHelper.projectNameToBinaryName(title);
@@ -178,6 +191,9 @@ public class IOSBundler implements IBundler {
 
         final boolean useArchive = true;
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         if (useArchive) {
             // Copy archive and game.projectc
             for (String name : Arrays.asList("game.projectc", "game.arci", "game.arcd", "game.dmanifest", "game.public.der")) {
@@ -187,6 +203,10 @@ public class IOSBundler implements IBundler {
             FileUtils.copyDirectory(buildDir, appDir);
             new File(buildDir, "game.arci").delete();
             new File(buildDir, "game.arcd").delete();
+        }
+
+        if(canceled.isCanceled()) {
+            return;
         }
 
         // Copy launch images
@@ -232,6 +252,9 @@ public class IOSBundler implements IBundler {
         List<String> applicationQueriesSchemes = new ArrayList<String>();
         List<String> urlSchemes = new ArrayList<String>();
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         String facebookAppId = projectProperties.getStringValue("facebook", "appid", null);
         if (facebookAppId != null) {
             urlSchemes.add("fb" + facebookAppId);
@@ -264,6 +287,10 @@ public class IOSBundler implements IBundler {
         properties.put("url-schemes", urlSchemes);
         properties.put("application-queries-schemes", applicationQueriesSchemes);
 
+        if(canceled.isCanceled()) {
+            return;
+        }
+
         List<String> orientationSupport = new ArrayList<String>();
         if(projectProperties.getBooleanValue("display", "dynamic_orientation", false)==false) {
             Integer displayWidth = projectProperties.getIntValue("display", "width", 960);
@@ -285,9 +312,15 @@ public class IOSBundler implements IBundler {
         helper.format(properties, "ios", "infoplist", new File(appDir, "Info.plist"));
         helper.copyIosIcons();
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Copy bundle resources into .app folder
         ExtenderUtil.writeResourcesToDirectory(bundleResources, appDir);
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Copy Provisioning Profile
         File provisioningProfileFile = new File(provisioningProfile);
         if (!provisioningProfileFile.exists()) {
@@ -295,11 +328,17 @@ public class IOSBundler implements IBundler {
         }
         FileUtils.copyFile(provisioningProfileFile, new File(appDir, "embedded.mobileprovision"));
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Create fat/universal binary
         File tmpFile = File.createTempFile("dmengine", "");
         tmpFile.deleteOnExit();
         String exe = tmpFile.getPath();
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Run lipo to add exeArmv7 + exeArm64 together into universal bin
         Result lipoResult = Exec.execResult( Bob.getExe(Platform.getHostPlatform(), "lipo"), "-create", exeArmv7.getAbsolutePath(), exeArm64.getAbsolutePath(), "-output", exe );
         if (lipoResult.ret == 0) {
@@ -309,6 +348,9 @@ public class IOSBundler implements IBundler {
             logger.log(Level.SEVERE, "Error executing lipo command:\n" + new String(lipoResult.stdOutErr));
         }
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Strip executable
         if( strip_executable )
         {
@@ -321,6 +363,9 @@ public class IOSBundler implements IBundler {
             }
         }
 
+        if(canceled.isCanceled()) {
+            return;
+        }
         // Copy Executable
         File destExecutable = new File(appDir, exeName);
         FileUtils.copyFile(new File(exe), destExecutable);
@@ -340,6 +385,9 @@ public class IOSBundler implements IBundler {
             File entitlementOut = File.createTempFile("entitlement", ".xcent");
             String customEntitlementsProperty = projectProperties.getStringValue("ios", "entitlements");
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             try {
                 XMLPropertyListConfiguration customEntitlements = new XMLPropertyListConfiguration();
                 XMLPropertyListConfiguration decodedProvision = new XMLPropertyListConfiguration();
@@ -378,6 +426,9 @@ public class IOSBundler implements IBundler {
                 throw new RuntimeException(e);
             }
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             ProcessBuilder processBuilder = new ProcessBuilder("codesign",
                     "-f", "-s", identity,
                     "--entitlements", entitlementOut.getAbsolutePath(),
@@ -385,30 +436,51 @@ public class IOSBundler implements IBundler {
             processBuilder.environment().put("EMBEDDED_PROFILE_NAME", "embedded.mobileprovision");
             processBuilder.environment().put("CODESIGN_ALLOCATE", Bob.getExe(Platform.getHostPlatform(), "codesign_allocate"));
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             Process process = processBuilder.start();
             logProcess(process);
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             // Package zip file
             File tmpZipDir = createTempDirectory();
             tmpZipDir.deleteOnExit();
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             // NOTE: We replaced the java zip file implementation(s) due to the fact that XCode didn't want
             // to import the resulting zip files.
             File payloadDir = new File(tmpZipDir, "Payload");
             payloadDir.mkdir();
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             processBuilder = new ProcessBuilder("cp", "-r", appDir.getAbsolutePath(), payloadDir.getAbsolutePath());
             process = processBuilder.start();
             logProcess(process);
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             File zipFile = new File(bundleDir, title + ".ipa");
             File zipFileTmp = new File(bundleDir, title + ".ipa.tmp");
             processBuilder = new ProcessBuilder("zip", "-qr", zipFileTmp.getAbsolutePath(), payloadDir.getName());
             processBuilder.directory(tmpZipDir);
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             process = processBuilder.start();
             logProcess(process);
 
+            if(canceled.isCanceled()) {
+                return;
+            }
             Files.move( Paths.get(zipFileTmp.getAbsolutePath()), Paths.get(zipFile.getAbsolutePath()), StandardCopyOption.ATOMIC_MOVE, StandardCopyOption.REPLACE_EXISTING);
             logger.log(Level.INFO, "Finished ipa: " + getFileDescription(zipFile));
         }
